Split linter tests out by file type
