+++ packages/analyticsdx-template-lint/test/unit/validator.test.ts
@@ -86,7 +86,8 @@ describe('FileTemplateValidator', () => {
       missingField: 'releaseInfo',
       lintWarning: {
         range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        message: 'App templates must have at least 1 dashboard, dataflow, externalFile, lens, or recipe specified',
+        message:
+          'App templates must have at least 1 dashboard, dataflow, dataTransform, externalFile, lens, or recipe specified',
         code: ERRORS.TMPL_APP_MISSING_OBJECTS
       }
     },
@@ -95,7 +96,8 @@ describe('FileTemplateValidator', () => {
       missingField: 'releaseInfo',
       lintWarning: {
         range: { start: { line: 1, character: 2 }, end: { line: 1, character: 23 } },
-        message: 'App templates must have at least 1 dashboard, dataflow, externalFile, lens, or recipe specified',
+        message:
+          'App templates must have at least 1 dashboard, dataflow, dataTransform, externalFile, lens, or recipe specified',
         code: ERRORS.TMPL_APP_MISSING_OBJECTS
       }
     },
